Rename SolverInterface to Participant and remove old API functions
